% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_cell_filter.R
\name{MxIF.cell.filter}
\alias{MxIF.cell.filter}
\title{MxIF Cell Filtering}
\usage{
MxIF.cell.filter(data, MinCellSize = 150, MaxCellSize = 2000,
  MinNucSize = 30, MaxNucSize = 1000, MinMemSize = 20, MinCytSize = 0)
}
\arguments{
\item{data}{data frame of cell data from Matlab based segmentation}

\item{MinCellSize}{minumum cell area (pixels), default=150}

\item{MinNucSize}{minumum nuclear area (pixels), default=30}

\item{MaxNucSize}{maximum nuclear area (pixels), default=1000}

\item{MinMemSize}{minimum membrane area (pixels), default=20}

\item{MinCytSize}{minimum cytoplasm area (pixels), default=0}

\item{MinCellSize}{maximum cell area (pixels), default=2000}
}
\description{
This function returns cells filtered for specified properties.
Eliot McKinley 9/15/17
}
\examples{
cell.filter(data)
cell.filter(data, MinCellSize=50, MaxCellSize=2500, MinNucSize=20, MaxNucSize=1500, MinMemSize=50, MinCytSize=10)
}
\keyword{MxIF}
\keyword{filter}
