% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_cell_filter_view.R
\name{MxIF.cell.filter.view}
\alias{MxIF.cell.filter.view}
\title{MxIF Cell Filter View}
\usage{
MxIF.cell.filter.view(data, LABELS, NOVLP)
}
\arguments{
\item{data}{data frame of cell data from Matlab based segmentation}

\item{LABELS}{tiff image of cell segmentation (ID=grayscale value)}

\item{NOVLP}{png image of the Non-overlap segmentation results}
}
\description{
This function outputs visulaization of cells filtered.
Eliot McKinley 9/25/17
}
\examples{
MxIF.tsne.plot(data, 3)
MxIF.tsne.plot(data, 3, colors="threecolor")
MxIF.tsne.plot(data, 3, colors=c("#000000","#98EF56"))
MxIF.tsne.plot(data, 3, colors=c("#000000","#98EF56"), sz=2, highQ=0.95, lowQ=0.05)
}
\keyword{MxIF}
\keyword{cell}
\keyword{filter}
