% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_cell_ID_view.R
\name{MxIF.cell.ID.view}
\alias{MxIF.cell.ID.view}
\title{MxIF Cell ID View}
\usage{
MxIF.cell.ID.view(AFRemoved, LABELS, IDs, alpha = 0.3)
}
\arguments{
\item{AFRemoved}{Image underlay for cell ID view}

\item{LABELS}{tiff image of cell segmentation (ID=grayscale value)}

\item{IDs}{array of cell IDs to overlay}

\item{alpha}{transparency or red overlay}
}
\description{
This function outputs red overlay of selected cells.
Eliot McKinley 9/27/17
}
\examples{
MxIF.cell.ID.view(AFRemoved, LABELS, c(100,200), alpha=.2)
}
\keyword{Ids}
\keyword{MxIF}
\keyword{cell}
\keyword{filter}
\keyword{viewer}
