% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF.death.star.plot.R
\name{MxIF.death.star.plot}
\alias{MxIF.death.star.plot}
\title{MxIF Create Death Star Plot following binary deconvolution}
\usage{
MxIF.death.star.plot(data, markers, file, color = "#0709b2", title = TRUE)
}
\arguments{
\item{data}{data frame of binary data (positive/negative) for each marker, e.g. from MxIF.deconvolution function}

\item{markers}{which markers to create graphs for}

\item{file}{string of where to save data}

\item{color}{hex denoted color for positive markers, default="#0709b2" - navy blue}

\item{title}{whether to include title with marker names for each ring, default=TRUE}
}
\description{
This function creates a circular "death star" plot from binary binned data and outputs counts for each group. Title of the plot denotes the marker corresponding to each concentric ring.
Based upon:
McKinley ET, Sui Y, Al-Kofahi Y, Millis BA, Tyska MJ, Roland JT, Santamaria-Pang A, Ohland CL, Jobin C, Franklin JL, Lau KS, Gerdes MJ, Coffey RJ. Optimized multiplex immunofluorescence single-cell analysis reveals tuft cell heterogeneity. JCI Insight. 2017 Jun 2; 2(11): e93487. 
https://insight.jci.org/articles/view/93487
}
\details{
Eliot McKinley 10/20/17
}
\examples{
MxIF.death.star.plot(data, markers, file, color="#0709b2", title=TRUE))
}
\keyword{MxIF}
\keyword{binary}
\keyword{circular}
\keyword{death}
\keyword{plots}
\keyword{star}
