% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_deconvolution.R
\name{MxIF.deconvoltion}
\alias{MxIF.deconvoltion}
\title{MxIF Model Based Deconvolution Threshold}
\usage{
MxIF.deconvoltion(data, markers, slidename, file, selectedMarker = "",
  quant = 0, plots = FALSE, cellIDcols = c(1:2))
}
\arguments{
\item{data}{data frame of cell data from Matlab based segmentation}

\item{markers}{which markers to run deconvolution processing on}

\item{slidename}{string of slidename for data output}

\item{file}{string of where to save data}

\item{selectedMarker}{string of which marker used to trim data by quantile, default =''}

\item{quant}{what quantile cutoff to use, defualt = ''}

\item{plots}{whether to output plots of steps, default =FALSE}

\item{celIDcols}{which columns to keep to ID cells for future processing, default= c(1,2) CellID and Pos}
}
\description{
This function uses model based deconvolution to threshold markers into two groups (high/low). Output is a table of cells with positive(1) or negative(0) values for each marker. Also writes table of threshold values. Default is for all data. If user inputs a selected marker and a quantile, operations will only be run on the cells for the top 1-quant quantile for that marker. 
Based upon:
McKinley ET, Sui Y, Al-Kofahi Y, Millis BA, Tyska MJ, Roland JT, Santamaria-Pang A, Ohland CL, Jobin C, Franklin JL, Lau KS, Gerdes MJ, Coffey RJ. Optimized multiplex immunofluorescence single-cell analysis reveals tuft cell heterogeneity. JCI Insight. 2017 Jun 2; 2(11): e93487. 
https://insight.jci.org/articles/view/93487
}
\details{
Eliot McKinley 10/18/17
}
\examples{
MxIF.deconvolution(data, markers, slidename, file, selectedMarker='Median_Cell_GFP', quant=0.8, plots=TRUE, cellIDcols=c(1,2,24))
}
\keyword{MxIF}
\keyword{deconvolution}
\keyword{threshold}
