% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_stat_cat.R
\name{MxIF.stat.cat}
\alias{MxIF.stat.cat}
\title{MxIF Stat Cat}
\usage{
MxIF.stat.cat(file)
}
\arguments{
\item{file}{location of folder containing stats files}
}
\description{
This function concatenates stats files into one CSV.
Eliot McKinley 9/15/17
}
\examples{
MxIF.stat.cat(file)
}
\keyword{MxIF}
\keyword{concatenate}
\keyword{stats}
