% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_tsne_animate.R
\name{MxIF.tsne.animate}
\alias{MxIF.tsne.animate}
\title{MxIF tSNE animate}
\usage{
MxIF.tsne.animate(data, markers, file, colors = "colorblind", sz = 1.5,
  highQ = 0.98, lowQ = 0.02)
}
\arguments{
\item{data}{data frame of cell data from Matlab based segmentation}

\item{markers}{columns to display}

\item{file}{where to save data}

\item{colors}{what colors to use, default is "colorblind" friendly, can also choose "threecolor" black-red-blue, or set your own}

\item{sz, }{what size dot to display, default=1.5}

\item{highQ}{max percentile default=.98}

\item{lowQ}{max percentile default=.02}
}
\description{
This function plots tSNE results as an animated gif. N.B. column names for plotting must be tSNE1 and tSNE2
Eliot McKinley 9/15/17
}
\examples{
MxIF.tsne.animate(posData, markers, file)
MxIF.tsne.animate(posData, markers, file, colors="threecolor")
MxIF.tsne.animate(posData, markers, file, colors=c("#000000","#98EF56"))
MxIF.tsne.animate(posData, markers, file, colors=c("#000000","#98EF56"), sz=2, highQ=0.95, lowQ=0.05)
}
\keyword{MxIF}
\keyword{tSNE}
