% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxIF_tSNE.R
\name{MxIF.tsne}
\alias{MxIF.tsne}
\title{MxIF tSNE}
\usage{
MxIF.tsne(data, markers, seed = 42, perplexity = 50, iterations = 1000,
  verbose = TRUE)
}
\arguments{
\item{data}{data frame of cell data from Matlab based segmentation}

\item{markers}{columns for each marker to run tSNE on}

\item{seed}{set seed for tSNE, default=42}

\item{perplexity, }{set perplexity for tSNE, default=50}

\item{iterations}{set iterations for tSNE , default=1000}

\item{verbose}{whether to output tSNE iteration steps to console, default=true}
}
\description{
This function runs 2D tSNE on a data set, user specifies columns for processing.
Eliot McKinley 9/15/17
}
\examples{
MxIF.tsne(data, c(1,2, 5:9))
MxIF.tsne(data, c(1:5, 6 ,8) seed=50, perplexity=30, iterations= 500, verbose=FALSE)

}
\keyword{MxIF}
\keyword{tSNE}
